rename color.ts -> color-ecs.ts
