+++ src/game/bullet.ts
@@ -1,7 +1,7 @@
 import { EM, EntityManager, Component, Entity } from "../entity-manager.js";
 import { quat, vec3 } from "../gl-matrix.js";
 import { FinishedDef } from "../build.js";
-import { ColorDef } from "../color.js";
+import { ColorDef } from "../color-ecs.js";
 import { RenderableConstructDef } from "../render/renderer-ecs.js";
 import { Position, PositionDef, RotationDef } from "../physics/transform.js";
 import { ColliderDef } from "../physics/collider.js";
