+++ src/game/cursor.ts
@@ -9,7 +9,7 @@ import {
   updateFrameFromPosRotScale,
 } from "../physics/transform.js";
 import { AssetsDef } from "./assets.js";
-import { ColorDef } from "../color.js";
+import { ColorDef } from "../color-ecs.js";
 import { assert } from "../test.js";
 import { CameraViewDef } from "../camera.js";
 import { vec2, vec3 } from "../gl-matrix.js";
