+++ src/game/game-hyperspace.ts
@@ -1,5 +1,5 @@
 import { CameraDef } from "../camera.js";
-import { ColorDef } from "../color.js";
+import { ColorDef } from "../color-ecs.js";
 import { EntityManager, EntityW } from "../entity-manager.js";
 import { PositionDef, RotationDef, ScaleDef } from "../physics/transform.js";
 import { RendererDef, RenderableConstructDef } from "../render/renderer-ecs.js";
