+++ src/game/game-sandbox.ts
@@ -4,7 +4,7 @@ import {
   setCameraFollowPosition,
   CameraDef,
 } from "../camera.js";
-import { ColorDef } from "../color.js";
+import { ColorDef } from "../color-ecs.js";
 import { DeletedDef } from "../delete.js";
 import { EM, EntityManager } from "../entity-manager.js";
 import { vec3, quat, mat4 } from "../gl-matrix.js";
