+++ src/game/modeler.ts
@@ -15,7 +15,7 @@ import { RenderableConstructDef } from "../render/renderer-ecs.js";
 import { tempVec3 } from "../temp-pool.js";
 import { vec3Dbg } from "../utils-3d.js";
 import { AssetsDef } from "./assets.js";
-import { ColorDef, TintsDef } from "../color.js";
+import { ColorDef, TintsDef } from "../color-ecs.js";
 import { drawLine } from "../utils-game.js";
 import { CameraView, CameraViewDef } from "../camera.js";
