+++ src/net/net-debug.ts
@@ -2,7 +2,7 @@ import { Component, EM, EntityManager } from "../entity-manager.js";
 import { vec3 } from "../gl-matrix.js";
 import { InputsDef } from "../inputs.js";
 import { RenderableDef } from "../render/renderer-ecs.js";
-import { clearTint, setTint, TintsDef } from "../color.js";
+import { clearTint, setTint, TintsDef } from "../color-ecs.js";
 import { AuthorityDef } from "./components.js";
 
 const NetDebugStateDef = EM.defineComponent("netDebugState", () => ({
