+++ src/physics/narrowphase.ts
@@ -1,6 +1,6 @@
 import { EntityManager, EntityW } from "../entity-manager.js";
 import { AssetsDef } from "../game/assets.js";
-import { ColorDef } from "../color.js";
+import { ColorDef } from "../color-ecs.js";
 import { LocalPlayerDef } from "../game/player.js";
 import { vec3 } from "../gl-matrix.js";
 import { RenderableConstructDef } from "../render/renderer-ecs.js";
