+++ src/physics/phys-debug.ts
@@ -1,7 +1,7 @@
 import { ColliderDef } from "./collider.js";
 import { EM, EntityManager } from "../entity-manager.js";
 import { AssetsDef, LocalMeshes } from "../game/assets.js";
-import { ColorDef } from "../color.js";
+import { ColorDef } from "../color-ecs.js";
 import { InputsDef } from "../inputs.js";
 import { mathMap } from "../math.js";
 import { cloneMesh, mapMeshPositions, RawMesh } from "../render/mesh.js";
